chore(wasm): rename wasm_runtime folder to runtime & remove unneeded files
