+++ packages/wasm/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod error;
-pub mod utils;
-pub mod wasm_runtime;
+pub mod runtime;
 pub mod file_reader;
 pub mod wasm_wrapper;
 pub mod wasm_package;
